+++ packages/cli/plugin-garfish/src/cli/index.ts
@@ -1,7 +1,6 @@
 import path from 'path';
 import { createRuntimeExportsUtils, PLUGIN_SCHEMAS } from '@modern-js/utils';
 import type { CliHookCallbacks, CliPlugin } from '@modern-js/core';
-import type WebpackChain from 'webpack-chain';
 import { logger } from '../util';
 import {
   getRuntimeConfig,
@@ -75,19 +74,17 @@ export default ({
             },
           },
           tools: {
+            devServer: {
+              headers: {
+                'Access-Control-Allow-Origin': '*',
+              },
+            },
             webpack: (
               webpackConfig: any,
-              {
-                chain,
-                webpack,
-                env = process.env.NODE_ENV || 'development',
-              }: { chain: WebpackChain; webpack: any; env: string },
+              { chain, webpack, env = process.env.NODE_ENV || 'development' },
             ) => {
               if (resolveOptions?.deploy?.microFrontend) {
                 chain.output.libraryTarget('umd');
-                chain.devServer.headers({
-                  'Access-Control-Allow-Origin': '*',
-                });
                 if (resolveOptions?.server?.port) {
                   chain.output.publicPath(
                     env === 'development'