+++ packages/solutions/app-tools/src/utils/types.ts
@@ -1,7 +1,12 @@
 export type DevOptions = {
   entry?: string[] | boolean;
+  apiOnly?: boolean;
 };
 
 export type BuildOptions = {
   analyze?: boolean;
+};
+
+export type StartOptions = {
+  apiOnly?: boolean;
 };