+++ packages/server/plugin-express/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @modern-js/plugin-express
 
+## 1.5.0
+
+### Minor Changes
+
+- d2d1d6b2: feat: support server config
+
+### Patch Changes
+
+- Updated dependencies [77ff9754]
+- Updated dependencies [d2d1d6b2]
+- Updated dependencies [07a4887e]
+- Updated dependencies [ea2ae711]
+- Updated dependencies [17d0cc46]
+- Updated dependencies [d2d1d6b2]
+  - @modern-js/utils@1.4.0
+  - @modern-js/types@1.4.0
+
 ## 1.4.4
 
 ### Patch Changes